+++ packages/translator/src/translate.ts
@@ -4,7 +4,10 @@ import { getAiTranslation } from './x'
 // Types
 import type { Config, Locale, Messages } from '@rosetta.js/types'
 
-type TranslateOptions = Pick<Config, 'projectLocale' | 'outLocales'> & {
+type TranslateOptions = Pick<
+  Config,
+  'projectLocale' | 'outLocales' | 'openaiApiKey'
+> & {
   messages: Messages
   locale: Locale
 }
@@ -12,14 +15,16 @@ type TranslateOptions = Pick<Config, 'projectLocale' | 'outLocales'> & {
 export async function translate({
   messages,
   projectLocale,
-  locale
+  locale,
+  openaiApiKey
 }: TranslateOptions) {
   const messagesJson = JSON.stringify(messages)
 
   const translation = await getAiTranslation({
     messagesJson,
     locale,
-    projectLocale
+    projectLocale,
+    openaiApiKey
   })
 
   return JSON.parse(translation)