+++ packages/types/src/config.ts
@@ -3,8 +3,9 @@ import type { Locale } from './locale'
 export type Config = {
   include: string[]
   exclude?: string[]
+  openaiApiKey: string
+  projectLocale: Locale
   outDir: string
   outLocales: Locale[]
-  projectLocale: Locale
   outExtension: '.ts' | '.js'
 }