+++ packages/dev/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @rosetta.js/dev
 
+## 0.1.2
+
+### Patch Changes
+
+- Enabled use personal OpenAI API Key
+- 07d9d25: Code improvements
+- Updated dependencies
+- Updated dependencies [07d9d25]
+  - @rosetta.js/translator@0.1.2
+  - @rosetta.js/node@0.1.2
+  - @rosetta.js/extractor@0.1.2
+  - @rosetta.js/logger@0.1.2
+  - @rosetta.js/utils@0.1.2
+
 ## 0.0.1
 
 ### Patch Changes