+++ packages/extractor/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @rosetta.js/extractor
 
+## 0.1.2
+
+### Patch Changes
+
+- Enabled use personal OpenAI API Key
+- 07d9d25: Code improvements
+- Updated dependencies
+  - @rosetta.js/node@0.1.2
+  - @rosetta.js/logger@0.1.2
+  - @rosetta.js/transformer@0.1.2
+
 ## 0.0.1
 
 ### Patch Changes