+++ packages/node/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rosetta.js/node
 
+## 0.1.2
+
+### Patch Changes
+
+- Enabled use personal OpenAI API Key
+- Updated dependencies
+  - @rosetta.js/types@0.1.2
+  - @rosetta.js/logger@0.1.2
+
 ## 0.0.1
 
 ### Patch Changes