+++ playground/vite/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rosetta.js/playground-vite
 
+## 0.1.2
+
+### Patch Changes
+
+- Enabled use personal OpenAI API Key
+- 07d9d25: Code improvements
+- Updated dependencies
+  - @rosetta.js/react@0.1.2
+
 ## 0.0.1
 
 ### Patch Changes