+++ website/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rosetta.js/website
 
+## 0.1.2
+
+### Patch Changes
+
+- Enabled use personal OpenAI API Key
+
 ## 0.1.1
 
 ### Patch Changes